Refactor transforms
