+++ lib/cli.js
@@ -16,6 +16,7 @@ var dir = require('node-dir');
 
 var BASE_DIR_SCHEMA = '../src/main/resources/schema/ans/';
 var BASE_DIR_TESTS = '../tests/fixtures/';
+var BASE_DIR_UPVERTERS = '../lib/upvert/';
 
 // Set default options
 var options = {
@@ -105,6 +106,19 @@ var commands = {
             }
           );
         },
+
+        function addUpverter(callback) {
+          console.log("Creating upverter stub...");
+          var source = path.join(path.dirname(module.filename), BASE_DIR_UPVERTERS, 'template.js');
+          var destination = path.join(path.dirname(module.filename), BASE_DIR_UPVERTERS, current_version + '.js');
+          ncp(source, destination, {
+            'transform': function transform(read, write) {
+              read
+                .pipe(replaceStream("NEXT_VERSION", new_version))
+                .pipe(write);
+            }}, callback);
+        },
+
         function appendVersion(callback) {
           console.log("Writing version history");
           var history = version.history;
@@ -166,8 +180,14 @@ var commands = {
           rimraf(tests_path, {}, callback);
         },
 
+        function deleteUpverter(callback) {
+          var upverter_path = path.join(path.dirname(module.filename), BASE_DIR_UPVERTERS, prev_version + '.js');
+          console.log("Deleting " + upverter_path + "...");
+          rimraf(upverter_path, {}, callback);
+        },
+
         function removeVersion(callback) {
-          console.log("Removing " + current_version + " from version history...");
+          console.log("Removing " + prev_version + " from version history...");
           var history = version.history;
           var removed = false;
           _.forEach(history, function(item, index) {
